Creates the IgnoreDuplicateTypes bridge.json setting.
The setting will turn an error into a warning when a same type is
extracted while loading nested assemblies (references of references).

If true: emits a warning when an already loaded type (full namespace path)
is found in an assembly being loaded. Skips loading that specific type,
yet continues to parse the assembly, so any non-overlapping type is
loaded from that assembly.

If false (default, and current Bridge behavior): throws an exception
when an already loaded type (full namespace path) is found in an
assembly being loaded. The Bridge translator process returns with error
and the build is assumed as failing.

This implements the feature described in issue bridgedotnet/Bridge#3997.