Fixes Bridge.Translator.Tests project reference to Dataflow.
The project was trying to reference the NuGet package albeit not having it
within its packages.config. Upon upgrading the package in other projects,
it would then break or reference the latest available in the system (which
happens to be located within msbuild installation paths).

Now a proper NuGet reference is in place, so that the reference is in sync
with Bridge.Translator's package.